# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PocketSync API
  description: |
    Consolidated personal finance manager API for the Nigerian market.

    **Production base URL:** `https://pocketsync.onrender.com`

    ## Authentication
    PocketSync uses **HttpOnly cookies** — not Bearer tokens. After `POST /api/v1/auth/login`,
    the server sets `accessToken` and `refreshToken` cookies. All protected routes require
    the browser/client to send cookies (`credentials: 'include'` in fetch, `withCredentials: true` in axios).

    ## Sign-up & password reset (email OTP)
    1. `POST /api/v1/auth/register` — creates account, sends 6-digit email OTP
    2. `POST /api/v1/auth/verify-otp` with `purpose: signup` — marks email verified
    3. `POST /api/v1/auth/login` — blocked until email is verified (`403`)
    4. Forgot password: `POST /api/v1/auth/forgot-password` → `POST /api/v1/auth/reset-password`

    OTP resend cooldown is **30 seconds**. Codes expire in **10 minutes**.
    Without `RESEND_API_KEY`, OTPs are logged to the server console; in development the code
    may also appear in the JSON response as `devOtp`.

    ## BVN onboarding (mocked — Phase 2)
    After email verification and login, complete onboarding:
    1. `GET /api/v1/onboarding/status` — route user to the correct step
    2. `POST /api/v1/onboarding/bvn/submit` — mock BVN + phone validation, sends SMS OTP
    3. `POST /api/v1/onboarding/bvn/verify-otp` — verifies phone, discovers mock accounts
...

...

    ## Amounts
...

...
servers:
...
paths:
...
  /api/v1/onboarding/status:
    get:
      tags: [Onboarding]
      summary: Get onboarding progress
      description: Use `currentStep` to route the user after login.
      operationId: getOnboardingStatus
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Onboarding status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/onboarding/bvn/submit:
    post:
      tags: [Onboarding]
      summary: Submit BVN and phone (mock validation)
      description: Accepts any valid 11-digit BVN and Nigerian phone. Sends mock SMS OTP.
      operationId: submitBvn
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BvnSubmitRequest'
            example:
              bvn: '22234567890'
              phone: '08012345678'
      responses:
        '200':
          description: BVN accepted, OTP sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BvnSubmitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: BVN already linked to another account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/onboarding/bvn/send-otp:
    post:
      tags: [Onboarding]
      summary: Resend phone OTP
      operationId: sendBvnOtp
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneOtpSentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Resend cooldown active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpCooldownError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/onboarding/bvn/verify-otp:
    post:
      tags: [Onboarding]
      summary: Verify phone OTP and discover accounts
      operationId: verifyBvnOtp
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneOtpVerifyRequest'
      responses:
        '200':
          description: Phone verified, mock accounts discovered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BvnVerifyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/onboarding/bvn/accounts:
    get:
      tags: [Onboarding]
      summary: List discovered accounts from BVN
      operationId: getDiscoveredAccounts
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Unlinked discovered accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveredAccountsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/onboarding/bvn/connect:
    post:
      tags: [Onboarding]
      summary: Connect selected discovered accounts
      operationId: connectDiscoveredAccounts
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BvnConnectRequest'
      responses:
        '201':
          description: Accounts linked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BvnConnectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: HttpOnly cookie set by POST /api/v1/auth/login
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Not authenticated or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorised — please log in
    Forbidden:
      description: Resource exists but not owned by user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          description: Present in development only
...
    OtpCooldownError:
      type: object
      properties:
        error:
          type: string
        resendAvailableIn:
          type: integer
...
    OnboardingStatusResponse:
      type: object
      properties:
        emailVerified:
          type: boolean
        bvnVerified:
          type: boolean
        phoneVerified:
          type: boolean
        currentStep:
          type: string
          enum: [bvn_entry, phone_otp, connect_accounts, complete]
        maskedPhone:
          type: string
          example: '+234 *** *** 5678'
        pendingAccounts:
          type: integer
        linkedAccounts:
          type: integer
        onboardingComplete:
          type: boolean
    BvnSubmitRequest:
      type: object
      required: [bvn, phone]
      properties:
        bvn:
          type: string
          pattern: '^\d{11}$'
        phone:
          type: string
          example: '08012345678'
    BvnSubmitResponse:
      type: object
      properties:
        message:
          type: string
        maskedPhone:
          type: string
        resendAvailableIn:
          type: integer
        devOtp:
          type: string
    PhoneOtpSentResponse:
      type: object
      properties:
        message:
          type: string
        maskedPhone:
          type: string
        resendAvailableIn:
          type: integer
        devOtp:
          type: string
    PhoneOtpVerifyRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          pattern: '^\d{6}$'
    BvnVerifyResponse:
      type: object
      properties:
        message:
          type: string
        bvnVerified:
          type: boolean
        phoneVerified:
          type: boolean
        discoveredAccounts:
          type: integer
        currentStep:
          type: string
          example: connect_accounts
    DiscoveredAccountsResponse:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/DiscoveredAccount'
    DiscoveredAccount:
      type: object
      properties:
        id:
          type: string
        institution:
          type: string
        maskedAccountNumber:
          type: string
        accountType:
          type: string
        balance:
          type: number
          description: NGN
        currency:
          type: string
          example: NGN
        holderName:
          type: string
    BvnConnectRequest:
      type: object
      required: [accountIds]
      properties:
        accountIds:
          type: array
          items:
            type: string
          minItems: 1
    BvnConnectResponse:
      type: object
      properties:
        message:
          type: string
        accounts:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              institution:
                type: string
              maskedAccountNumber:
                type: string
              balance:
                type: number
              accountType:
                type: string
        onboardingComplete:
          type: boolean
...